Changed double variables to float for division

+++ src/CurrencyConverter.java
@@ -1,6 +1,6 @@
 public interface CurrencyConverter {
-    public double convertToDollars(double amount);
-    public double convertToSterling(double amount);
-    public double convertToEuros(double amount);
+    public float convertToDollars(float amount);
+    public float convertToSterling(float amount);
+    public float convertToEuros(float amount);
 
 }
